requirements.txt
aiohttp==2.3.9
async-timeout==2.0.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
greenlet==0.4.12
idna==2.6
multidict==4.0.0
python-dotenv==0.7.1
redis==2.10.6
requests==2.18.4
rq==0.10.0
urllib3==1.22
yarl==1.1.0
